+++ Examples/minimum_user_for_s2e/src/src_user/Settings/TlmCmd/common_tlm_cmd_packet_define.h
@@ -5,7 +5,7 @@
 #ifndef COMMON_TLM_CMD_PACKET_DEFINE_H_
 #define COMMON_TLM_CMD_PACKET_DEFINE_H_
 
-#include "../../TlmCmd/Ccsds/TCPacket.h"
+#include <src_core/TlmCmd/Ccsds/TCPacket.h>
 
 // 使用するパケット構造体をCTCP型として定義する
 typedef TCP CTCP;
