+++ Examples/minimum_user_for_s2e/src/src_user/TlmCmd/Ccsds/m_pdu.c
@@ -3,7 +3,7 @@
 
 #include "string.h" // for memcpy
 
-#include "TCPacket.h"
+#include <src_core/TlmCmd/Ccsds/TCPacket.h>
 
 static void M_PDU_clear_spare_(M_PDU* m_pdu);
