+++ Examples/minimum_user_for_s2e/src/src_user/TlmCmd/common_tlm_cmd_packet.c
@@ -8,7 +8,7 @@
  */
 
 #include <string.h> // for memcpy
-#include "./Ccsds/TCPacket.h"
+#include <src_core/TlmCmd/Ccsds/TCPacket.h>
 #include <src_core/TlmCmd/common_tlm_cmd_packet.h>
 #include <src_core/Library/endian_memcpy.h>
