+++ TlmCmd/Ccsds/TCPacket.h
@@ -3,13 +3,17 @@
 
 #include <stddef.h> // for size_t
 
-#include "../../Library/stdint.h"
+#include <src_user/Library/stdint.h>
 
 #define TCP_MAX_LEN          (432u)
 #define TCP_PRM_HDR_LEN        (6u)
 #define TCP_TLM_2ND_HDR_LEN    (7u)
 #define TCP_CMD_2ND_HDR_LEN    (1u)
 #define TCP_CMD_USER_HDR_LEN   (8u)
+
+// TCP_MAX_LEN を再定義
+// TCP_APID, TCP_CMD_DEST_TYPE を定義する
+#include <src_user/Settings/TlmCmd/Ccsds/tcpacket_params.h>
 
 
 /**
@@ -53,25 +57,6 @@ typedef enum
   TCP_2ND_HDR_ABSENT  = 0, // 0b: Secondary Header Absent
   TCP_2ND_HDR_PRESENT = 1  // 1b: Secondary Header Present
 } TCP_2ND_HDR_FLAG;
-
-/**
- * @enum   TCP_APID
- * @brief  Application Process ID
- * @note   GSTOS の蓄積データの拡張子につく数字
- * @note   11bit
- * @note   CTCP_DEST_ID として typedef する
- */
-typedef enum
-{
-  TCP_APID_MOBC_CMD = 0x210,         // 01000010000b:
-  TCP_APID_AOBC_CMD = 0x211,         // 01000010001b:
-  TCP_APID_TOBC_CMD = 0x212,         // 01000010010b:
-  TCP_APID_TCAL_TLM = 0x410,         // 10000010000b: APID for TIME CARIBLATION TLM
-  TCP_APID_MIS_TLM  = 0x510,         // 10100010000b: APID for MIS TLM
-  TCP_APID_DUMP_TLM = 0x710,         // 11100010000b: APID for DUMP TLM
-  TCP_APID_FILL_PKT = 0x7ff,         // 11111111111b: APID for FILL PACKET
-  TCP_APID_UNKNOWN
-} TCP_APID;
 
 /**
  * @enum   TCP_SEQ_FLAG
@@ -132,23 +117,6 @@ typedef enum
   TCP_CMD_EXEC_TYPE_UTL = 0x04, // 04: Unixtime Timeline Command
   TCP_CMD_EXEC_TYPE_UNKNOWN
 } TCP_CMD_EXEC_TYPE;
-
-/**
- * @enum   TCP_CMD_DEST_TYPE
- * @brief  コマンドの解釈の宛先を規定
- * @note   TO_ME: 自分自身 → 自分自身のTLCやBCとして解釈
- * @note   TO_*:  転送先のTLやBCとして解釈（GSから来たコマンドを自身のキューにいれない）
- * @note   0x*0を想定
- * @note   下位4bitは TCP_CMD_EXEC_TYPE と CCP_EXEC_TYPE に
- */
-typedef enum
-{
-  TCP_CMD_DEST_TYPE_TO_ME     = 0x00,
-  TCP_CMD_DEST_TYPE_TO_MOBC   = 0x10,
-  TCP_CMD_DEST_TYPE_TO_AOBC   = 0x20,
-  TCP_CMD_DEST_TYPE_TO_TOBC   = 0x30,
-  TCP_CMD_DEST_TYPE_TO_UNKOWN = 0x40
-} TCP_CMD_DEST_TYPE;
 
 
 // *******************
